details in description
Got the player animations working
Dash works
Player Hit works
Player Death works
----
The other animations are in:
Wake up
Sleeping

---
Got the Koi fish working, not sure what happened last time, went back to the previous git push and working fine. They are in the scene, swimming, they just need to be set up.

--
Tomorrow we can work on some sounds, music transition, colliders and the rest of the cinematics.
There is something wrong with the player death, I have it set so that it disables the samurai movement but for some reason it just keeps doing this.
But the animation works, and is working as intended.